% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octave.connect.R
\name{octave_connect}
\alias{octave_connect}
\title{Knitr Engine That Passes R Arrays to Octave}
\usage{
octave_connect(octave_path)
}
\arguments{
\item{octave_path}{This should be the path to your local Octave binary or batch file.}
}
\description{
It is important to provide the engine with the right path to the Octave binary.
On my Windows 10 machine with Octave 6.1.0, I use:
"C:/Users/zander/AppData/Local/Programs/GNU Octave/Octave-6.1.0/mingw64/bin/octave"
}
